move sql queries to adapter repos instead of ORM and adjust them;
