+++ src/PrestaShopBundle/Resources/config/services/bundle/repository.yml
@@ -129,20 +129,6 @@ services:
       - "@doctrine.dbal.default_connection"
       - "%database_prefix%"
 
-  PrestaShopBundle\Entity\Repository\AttributeGroupRepository:
-    factory: [ '@doctrine.orm.default_entity_manager', getRepository ]
-    arguments:
-      - PrestaShopBundle\Entity\AttributeGroup
-
-  prestashop.core.admin.attribute_group.repository:
-    alias: PrestaShopBundle\Entity\Repository\AttributeGroupRepository
-    deprecated: ~
-
-  PrestaShopBundle\Entity\Repository\AttributeRepository:
-    factory: [ '@doctrine.orm.default_entity_manager', getRepository ]
-    arguments:
-      - PrestaShopBundle\Entity\Attribute
-
   prestashop.core.admin.feature_flag.repository:
     class: PrestaShopBundle\Entity\Repository\FeatureFlagRepository
     factory: [ '@doctrine.orm.default_entity_manager', getRepository ]
